mqtt-python/requirements.txt
paho-mqtt
pymongo
redis
